Split spherical harmonics into multiple textures when necessary
