+++ src/splatmesh/SplatGeometry.js
@@ -1,0 +1,38 @@
+import * as THREE from 'three';
+
+export class SplatGeometry {
+
+    /**
+     * Build the Three.js geometry that will be used to render the splats. The geometry is instanced and is made up of
+     * vertices for a single quad as well as an attribute buffer for the splat indexes.
+     * @param {number} maxSplatCount The maximum number of splats that the geometry will need to accomodate
+     * @return {THREE.InstancedBufferGeometry}
+     */
+    static build(maxSplatCount) {
+
+        const baseGeometry = new THREE.BufferGeometry();
+        baseGeometry.setIndex([0, 1, 2, 0, 2, 3]);
+
+        // Vertices for the instanced quad
+        const positionsArray = new Float32Array(4 * 3);
+        const positions = new THREE.BufferAttribute(positionsArray, 3);
+        baseGeometry.setAttribute('position', positions);
+        positions.setXYZ(0, -1.0, -1.0, 0.0);
+        positions.setXYZ(1, -1.0, 1.0, 0.0);
+        positions.setXYZ(2, 1.0, 1.0, 0.0);
+        positions.setXYZ(3, 1.0, -1.0, 0.0);
+        positions.needsUpdate = true;
+
+        const geometry = new THREE.InstancedBufferGeometry().copy(baseGeometry);
+
+        // Splat index buffer
+        const splatIndexArray = new Uint32Array(maxSplatCount);
+        const splatIndexes = new THREE.InstancedBufferAttribute(splatIndexArray, 1, false);
+        splatIndexes.setUsage(THREE.DynamicDrawUsage);
+        geometry.setAttribute('splatIndex', splatIndexes);
+
+        geometry.instanceCount = 0;
+
+        return geometry;
+    }
+}
