+++ src/splatmesh/SplatMesh.js
@@ -1,13 +1,15 @@
 import * as THREE from 'three';
-import { SplatScene } from './SplatScene.js';
-import { SplatTree } from './splattree/SplatTree.js';
-import { WebGLExtensions } from './three-shim/WebGLExtensions.js';
-import { WebGLCapabilities } from './three-shim/WebGLCapabilities.js';
-import { uintEncodedFloat, rgbaArrayToInteger } from './Util.js';
-import { Constants } from './Constants.js';
-import { SceneRevealMode } from './SceneRevealMode.js';
-import { LogLevel } from './LogLevel.js';
-import { getSphericalHarmonicsComponentCountForDegree } from './Util.js';
+import { SplatMaterial } from './SplatMaterial.js';
+import { SplatGeometry } from './SplatGeometry.js';
+import { SplatScene } from '../SplatScene.js';
+import { SplatTree } from '../splattree/SplatTree.js';
+import { WebGLExtensions } from '../three-shim/WebGLExtensions.js';
+import { WebGLCapabilities } from '../three-shim/WebGLCapabilities.js';
+import { uintEncodedFloat, rgbaArrayToInteger } from '../Util.js';
+import { Constants } from '../Constants.js';
+import { SceneRevealMode } from '../SceneRevealMode.js';
+import { LogLevel } from '../LogLevel.js';
+import { getSphericalHarmonicsComponentCountForDegree } from '../Util.js';
 
 const dummyGeometry = new THREE.BufferGeometry();
 const dummyMaterial = new THREE.MeshBasicMaterial();
@@ -108,568 +110,6 @@ export class SplatMesh extends THREE.Mesh {
         this.disposed = false;
         this.lastRenderer = null;
         this.visible = false;
-    }
-
-    /**
-     * Build the Three.js material that is used to render the splats.
-     * @param {number} dynamicMode If true, it means the scene geometry represented by this splat mesh is not stationary or
-     *                             that the splat count might change
-     * @param {boolean} antialiased If true, calculate compensation factor to deal with gaussians being rendered at a significantly
-     *                              different resolution than that of their training
-     * @param {number} maxScreenSpaceSplatSize The maximum clip space splat size
-     * @param {number} splatScale Value by which all splats are scaled in screen-space (default is 1.0)
-     * @param {number} pointCloudModeEnabled Render all splats as screen-space circles
-     * @param {number} maxSphericalHarmonicsDegree Degree of spherical harmonics to utilize in rendering splats
-     * @return {THREE.ShaderMaterial}
-     */
-    static buildMaterial(dynamicMode = false, antialiased = false, maxScreenSpaceSplatSize = 2048,
-                         splatScale = 1.0, pointCloudModeEnabled = false, maxSphericalHarmonicsDegree = 0) {
-
-        // Contains the code to project 3D covariance to 2D and from there calculate the quad (using the eigen vectors of the
-        // 2D covariance) that is ultimately rasterized
-        let vertexShaderSource = `
-            precision highp float;
-            #include <common>
-
-            attribute uint splatIndex;
-
-            uniform highp sampler2D covariancesTexture;
-            uniform highp usampler2D centersColorsTexture;
-            uniform highp sampler2D sphericalHarmonicsTexture;`;
-
-        if (dynamicMode) {
-            vertexShaderSource += `
-                uniform highp usampler2D transformIndexesTexture;
-                uniform highp mat4 transforms[${Constants.MaxScenes}];
-                uniform vec2 transformIndexesTextureSize;
-            `;
-        }
-
-        vertexShaderSource += `
-            uniform vec2 focal;
-            uniform float orthoZoom;
-            uniform int orthographicMode;
-            uniform int pointCloudModeEnabled;
-            uniform float inverseFocalAdjustment;
-            uniform vec2 viewport;
-            uniform vec2 basisViewport;
-            uniform vec2 covariancesTextureSize;
-            uniform vec2 centersColorsTextureSize;
-            uniform int sphericalHarmonicsDegree;
-            uniform vec2 sphericalHarmonicsTextureSize;
-            uniform int sphericalHarmonics8BitMode;
-            uniform float visibleRegionRadius;
-            uniform float visibleRegionFadeStartRadius;
-            uniform float firstRenderTime;
-            uniform float currentTime;
-            uniform int fadeInComplete;
-            uniform vec3 sceneCenter;
-            uniform float splatScale;
-
-            varying vec4 vColor;
-            varying vec2 vUv;
-
-            varying vec2 vPosition;
-
-            const float sqrt8 = sqrt(8.0);
-            const float minAlpha = 1.0 / 255.0;
-
-            const vec4 encodeNorm4 = vec4(1.0 / 255.0, 1.0 / 255.0, 1.0 / 255.0, 1.0 / 255.0);
-            const uvec4 mask4 = uvec4(uint(0x000000FF), uint(0x0000FF00), uint(0x00FF0000), uint(0xFF000000));
-            const uvec4 shift4 = uvec4(0, 8, 16, 24);
-            vec4 uintToRGBAVec (uint u) {
-               uvec4 urgba = mask4 & u;
-               urgba = urgba >> shift4;
-               vec4 rgba = vec4(urgba) * encodeNorm4;
-               return rgba;
-            }
-
-            vec2 getDataUV(in int stride, in int offset, in vec2 dimensions) {
-                vec2 samplerUV = vec2(0.0, 0.0);
-                float d = float(splatIndex * uint(stride) + uint(offset)) / dimensions.x;
-                samplerUV.y = float(floor(d)) / dimensions.y;
-                samplerUV.x = fract(d);
-                return samplerUV;
-            }
-
-            vec2 getDataUVF(in uint sIndex, in float stride, in uint offset, in vec2 dimensions) {
-                vec2 samplerUV = vec2(0.0, 0.0);
-                float d = float(uint(float(sIndex) * stride) + offset) / dimensions.x;
-                samplerUV.y = float(floor(d)) / dimensions.y;
-                samplerUV.x = fract(d);
-                return samplerUV;
-            }
-
-            const float SH_C1 = 0.4886025119029199f;
-            const float[5] SH_C2 = float[](1.0925484, -1.0925484, 0.3153916, -1.0925484, 0.5462742);
-
-            const float SphericalHarmonics8BitCompressionRange = ${Constants.SphericalHarmonics8BitCompressionRange.toFixed(1)};
-            const float SphericalHarmonics8BitCompressionHalfRange = SphericalHarmonics8BitCompressionRange / 2.0;
-            const vec3 vec8BitSHShift = vec3(SphericalHarmonics8BitCompressionHalfRange);
-
-            void main () {
-
-                uint oddOffset = splatIndex & uint(0x00000001);
-                uint doubleOddOffset = oddOffset * uint(2);
-                bool isEven = oddOffset == uint(0);
-                uint nearestEvenIndex = splatIndex - oddOffset;
-                float fOddOffset = float(oddOffset);
-
-                uvec4 sampledCenterColor = texture(centersColorsTexture, getDataUV(1, 0, centersColorsTextureSize));
-                vec3 splatCenter = uintBitsToFloat(uvec3(sampledCenterColor.gba));`;
-
-            if (dynamicMode) {
-                vertexShaderSource += `
-                    uint transformIndex = texture(transformIndexesTexture, getDataUV(1, 0, transformIndexesTextureSize)).r;
-                    mat4 transform = transforms[transformIndex];
-                    mat4 transformModelViewMatrix = modelViewMatrix * transform;
-                `;
-            } else {
-                vertexShaderSource += `mat4 transformModelViewMatrix = modelViewMatrix;`;
-            }
-
-            vertexShaderSource += `
-                vec4 viewCenter = transformModelViewMatrix * vec4(splatCenter, 1.0);
-
-                vec4 clipCenter = projectionMatrix * viewCenter;
-
-                float clip = 1.2 * clipCenter.w;
-                if (clipCenter.z < -clip || clipCenter.x < -clip || clipCenter.x > clip || clipCenter.y < -clip || clipCenter.y > clip) {
-                    gl_Position = vec4(0.0, 0.0, 2.0, 1.0);
-                    return;
-                }
-
-                vPosition = position.xy;
-                vColor = uintToRGBAVec(sampledCenterColor.r);
-            `;
-
-            if (maxSphericalHarmonicsDegree >= 1) {
-
-                vertexShaderSource += `   
-                if (sphericalHarmonicsDegree >= 1) {
-                `;
-
-                if (dynamicMode) {
-                    vertexShaderSource += `
-                        mat4 mTransform = modelMatrix * transform;
-                        vec3 worldViewDir = normalize(splatCenter - vec3(inverse(mTransform) * vec4(cameraPosition, 1.0)));
-                    `;
-                } else {
-                    vertexShaderSource += `
-                        vec3 worldViewDir = normalize(splatCenter - cameraPosition);
-                    `;
-                }
-
-                if (maxSphericalHarmonicsDegree >= 2) {
-                    vertexShaderSource += `
-                        vec4 sampledSH0123 = texture(sphericalHarmonicsTexture, getDataUV(6, 0, sphericalHarmonicsTextureSize));
-                        vec4 sampledSH4567 = texture(sphericalHarmonicsTexture, getDataUV(6, 1, sphericalHarmonicsTextureSize));
-                        vec4 sampledSH891011 = texture(sphericalHarmonicsTexture, getDataUV(6, 2, sphericalHarmonicsTextureSize));
-                        vec3 sh1 = sampledSH0123.rgb;
-                        vec3 sh2 = vec3(sampledSH0123.a, sampledSH4567.rg);
-                        vec3 sh3 = vec3(sampledSH4567.ba, sampledSH891011.r);
-                    `;
-                } else {
-                    vertexShaderSource += `
-                        vec2 shUV = getDataUVF(nearestEvenIndex, 2.5, doubleOddOffset, sphericalHarmonicsTextureSize);
-                        vec4 sampledSH0123 = texture(sphericalHarmonicsTexture, shUV);
-                        shUV = getDataUVF(nearestEvenIndex, 2.5, doubleOddOffset + uint(1), sphericalHarmonicsTextureSize);
-                        vec4 sampledSH4567 = texture(sphericalHarmonicsTexture, shUV);
-                        shUV = getDataUVF(nearestEvenIndex, 2.5, doubleOddOffset + uint(2), sphericalHarmonicsTextureSize);
-                        vec4 sampledSH891011 = texture(sphericalHarmonicsTexture, shUV);
-
-                        vec3 sh1 = vec3(sampledSH0123.rgb) * (1.0 - fOddOffset) + vec3(sampledSH0123.ba, sampledSH4567.r) * fOddOffset;
-                        vec3 sh2 = vec3(sampledSH0123.a, sampledSH4567.rg) * (1.0 - fOddOffset) + vec3(sampledSH4567.gba) * fOddOffset;
-                        vec3 sh3 = vec3(sampledSH4567.ba, sampledSH891011.r) * (1.0 - fOddOffset) + vec3(sampledSH891011.rgb) * fOddOffset;
-                    `;
-                }
-
-                vertexShaderSource += `
-                        if (sphericalHarmonics8BitMode == 1) {
-                            sh1 = sh1 * SphericalHarmonics8BitCompressionRange - vec8BitSHShift;
-                            sh2 = sh2 * SphericalHarmonics8BitCompressionRange - vec8BitSHShift;
-                            sh3 = sh3 * SphericalHarmonics8BitCompressionRange - vec8BitSHShift;
-                        }
-                        float x = worldViewDir.x;
-                        float y = worldViewDir.y;
-                        float z = worldViewDir.z;
-                        vColor.rgb += SH_C1 * (-sh1 * y + sh2 * z - sh3 * x);
-                `;
-
-                if (maxSphericalHarmonicsDegree >= 2) {
-
-                    vertexShaderSource += `
-                        if (sphericalHarmonicsDegree >= 2) {
-                            float xx = x * x;
-                            float yy = y * y;
-                            float zz = z * z;
-                            float xy = x * y;
-                            float yz = y * z;
-                            float xz = x * z;
-
-                            vec4 sampledSH12131415 = texture(sphericalHarmonicsTexture, getDataUV(6, 3, sphericalHarmonicsTextureSize));
-                            vec4 sampledSH16171819 = texture(sphericalHarmonicsTexture, getDataUV(6, 4, sphericalHarmonicsTextureSize));
-                            vec4 sampledSH20212223 = texture(sphericalHarmonicsTexture, getDataUV(6, 5, sphericalHarmonicsTextureSize));
-
-                            vec3 sh4 = sampledSH891011.gba;
-                            vec3 sh5 = sampledSH12131415.rgb;
-                            vec3 sh6 = vec3(sampledSH12131415.a, sampledSH16171819.rg);
-                            vec3 sh7 = vec3(sampledSH16171819.ba, sampledSH20212223.r);
-                            vec3 sh8 = sampledSH20212223.gba;
-
-                            if (sphericalHarmonics8BitMode == 1) {
-                                sh4 = sh4 * SphericalHarmonics8BitCompressionRange - vec8BitSHShift;
-                                sh5 = sh5 * SphericalHarmonics8BitCompressionRange - vec8BitSHShift;
-                                sh6 = sh6 * SphericalHarmonics8BitCompressionRange - vec8BitSHShift;
-                                sh7 = sh7 * SphericalHarmonics8BitCompressionRange - vec8BitSHShift;
-                                sh8 = sh8 * SphericalHarmonics8BitCompressionRange - vec8BitSHShift;
-                            }
-
-                            vColor.rgb +=
-                                (SH_C2[0] * xy) * sh4 +
-                                (SH_C2[1] * yz) * sh5 +
-                                (SH_C2[2] * (2.0 * zz - xx - yy)) * sh6 +
-                                (SH_C2[3] * xz) * sh7 +
-                                (SH_C2[4] * (xx - yy)) * sh8;
-                        }
-                    `;
-                }
-
-                vertexShaderSource += `
-               
-                    vColor.rgb = clamp(vColor.rgb, vec3(0.), vec3(1.));
-
-                }
-
-                `;
-            }
-
-            vertexShaderSource += `
-
-                vec4 sampledCovarianceA = texture(covariancesTexture,
-                                                  getDataUVF(nearestEvenIndex, 1.5, oddOffset, covariancesTextureSize));
-                vec4 sampledCovarianceB = texture(covariancesTexture,
-                                                  getDataUVF(nearestEvenIndex, 1.5, oddOffset + uint(1), covariancesTextureSize));
-
-                vec3 cov3D_M11_M12_M13 = vec3(sampledCovarianceA.rgb) * (1.0 - fOddOffset) +
-                                         vec3(sampledCovarianceA.ba, sampledCovarianceB.r) * fOddOffset;
-                vec3 cov3D_M22_M23_M33 = vec3(sampledCovarianceA.a, sampledCovarianceB.rg) * (1.0 - fOddOffset) +
-                                         vec3(sampledCovarianceB.gba) * fOddOffset;
-
-                // Construct the 3D covariance matrix
-                mat3 Vrk = mat3(
-                    cov3D_M11_M12_M13.x, cov3D_M11_M12_M13.y, cov3D_M11_M12_M13.z,
-                    cov3D_M11_M12_M13.y, cov3D_M22_M23_M33.x, cov3D_M22_M23_M33.y,
-                    cov3D_M11_M12_M13.z, cov3D_M22_M23_M33.y, cov3D_M22_M23_M33.z
-                );
-
-                mat3 J;
-                if (orthographicMode == 1) {
-                    // Since the projection is linear, we don't need an approximation
-                    J = transpose(mat3(orthoZoom, 0.0, 0.0,
-                                       0.0, orthoZoom, 0.0,
-                                       0.0, 0.0, 0.0));
-                } else {
-                    // Construct the Jacobian of the affine approximation of the projection matrix. It will be used to transform the
-                    // 3D covariance matrix instead of using the actual projection matrix because that transformation would
-                    // require a non-linear component (perspective division) which would yield a non-gaussian result.
-                    float s = 1.0 / (viewCenter.z * viewCenter.z);
-                    J = mat3(
-                        focal.x / viewCenter.z, 0., -(focal.x * viewCenter.x) * s,
-                        0., focal.y / viewCenter.z, -(focal.y * viewCenter.y) * s,
-                        0., 0., 0.
-                    );
-                }
-
-                // Concatenate the projection approximation with the model-view transformation
-                mat3 W = transpose(mat3(transformModelViewMatrix));
-                mat3 T = W * J;
-
-                // Transform the 3D covariance matrix (Vrk) to compute the 2D covariance matrix
-                mat3 cov2Dm = transpose(T) * Vrk * T;
-                `;
-
-            if (antialiased) {
-                vertexShaderSource += `
-                    float detOrig = cov2Dm[0][0] * cov2Dm[1][1] - cov2Dm[0][1] * cov2Dm[0][1];
-                    cov2Dm[0][0] += 0.3;
-                    cov2Dm[1][1] += 0.3;
-                    float detBlur = cov2Dm[0][0] * cov2Dm[1][1] - cov2Dm[0][1] * cov2Dm[0][1];
-                    float compensation = sqrt(max(detOrig / detBlur, 0.0));
-                `;
-            } else {
-                vertexShaderSource += `
-                    cov2Dm[0][0] += 0.3;
-                    cov2Dm[1][1] += 0.3;
-                    float compensation = 1.0;
-                `;
-            }
-
-            vertexShaderSource += `
-
-                vColor.a *= compensation;
-
-                if (vColor.a < minAlpha) return;
-
-                // We are interested in the upper-left 2x2 portion of the projected 3D covariance matrix because
-                // we only care about the X and Y values. We want the X-diagonal, cov2Dm[0][0],
-                // the Y-diagonal, cov2Dm[1][1], and the correlation between the two cov2Dm[0][1]. We don't
-                // need cov2Dm[1][0] because it is a symetric matrix.
-                vec3 cov2Dv = vec3(cov2Dm[0][0], cov2Dm[0][1], cov2Dm[1][1]);
-
-                vec3 ndcCenter = clipCenter.xyz / clipCenter.w;
-
-                // We now need to solve for the eigen-values and eigen vectors of the 2D covariance matrix
-                // so that we can determine the 2D basis for the splat. This is done using the method described
-                // here: https://people.math.harvard.edu/~knill/teaching/math21b2004/exhibits/2dmatrices/index.html
-                // After calculating the eigen-values and eigen-vectors, we calculate the basis for rendering the splat
-                // by normalizing the eigen-vectors and then multiplying them by (sqrt(8) * eigen-value), which is
-                // equal to scaling them by sqrt(8) standard deviations.
-                //
-                // This is a different approach than in the original work at INRIA. In that work they compute the
-                // max extents of the projected splat in screen space to form a screen-space aligned bounding rectangle
-                // which forms the geometry that is actually rasterized. The dimensions of that bounding box are 3.0
-                // times the maximum eigen-value, or 3 standard deviations. They then use the inverse 2D covariance
-                // matrix (called 'conic') in the CUDA rendering thread to determine fragment opacity by calculating the
-                // full gaussian: exp(-0.5 * (X - mean) * conic * (X - mean)) * splat opacity
-                float a = cov2Dv.x;
-                float d = cov2Dv.z;
-                float b = cov2Dv.y;
-                float D = a * d - b * b;
-                float trace = a + d;
-                float traceOver2 = 0.5 * trace;
-                float term2 = sqrt(max(0.1f, traceOver2 * traceOver2 - D));
-                float eigenValue1 = traceOver2 + term2;
-                float eigenValue2 = traceOver2 - term2;
-
-                if (pointCloudModeEnabled == 1) {
-                    eigenValue1 = eigenValue2 = 0.2;
-                }
-
-                if (eigenValue2 <= 0.0) return;
-
-                vec2 eigenVector1 = normalize(vec2(b, eigenValue1 - a));
-                // since the eigen vectors are orthogonal, we derive the second one from the first
-                vec2 eigenVector2 = vec2(eigenVector1.y, -eigenVector1.x);
-
-                // We use sqrt(8) standard deviations instead of 3 to eliminate more of the splat with a very low opacity.
-                vec2 basisVector1 = eigenVector1 * splatScale * min(sqrt8 * sqrt(eigenValue1), ${parseInt(maxScreenSpaceSplatSize)}.0);
-                vec2 basisVector2 = eigenVector2 * splatScale * min(sqrt8 * sqrt(eigenValue2), ${parseInt(maxScreenSpaceSplatSize)}.0);
-
-                if (fadeInComplete == 0) {
-                    float opacityAdjust = 1.0;
-                    float centerDist = length(splatCenter - sceneCenter);
-                    float renderTime = max(currentTime - firstRenderTime, 0.0);
-
-                    float fadeDistance = 0.75;
-                    float distanceLoadFadeInFactor = step(visibleRegionFadeStartRadius, centerDist);
-                    distanceLoadFadeInFactor = (1.0 - distanceLoadFadeInFactor) +
-                                               (1.0 - clamp((centerDist - visibleRegionFadeStartRadius) / fadeDistance, 0.0, 1.0)) *
-                                               distanceLoadFadeInFactor;
-                    opacityAdjust *= distanceLoadFadeInFactor;
-                    vColor.a *= opacityAdjust;
-                }
-
-                vec2 ndcOffset = vec2(vPosition.x * basisVector1 + vPosition.y * basisVector2) *
-                                 basisViewport * 2.0 * inverseFocalAdjustment;
-
-                vec4 quadPos = vec4(ndcCenter.xy + ndcOffset, ndcCenter.z, 1.0);
-                gl_Position = quadPos;
-
-                // Scale the position data we send to the fragment shader
-                vPosition *= sqrt8;
-            }`;
-
-        const fragmentShaderSource = `
-            precision highp float;
-            #include <common>
- 
-            uniform vec3 debugColor;
-
-            varying vec4 vColor;
-            varying vec2 vUv;
-
-            varying vec2 vPosition;
-
-            void main () {
-                // Compute the positional squared distance from the center of the splat to the current fragment.
-                float A = dot(vPosition, vPosition);
-                // Since the positional data in vPosition has been scaled by sqrt(8), the squared result will be
-                // scaled by a factor of 8. If the squared result is larger than 8, it means it is outside the ellipse
-                // defined by the rectangle formed by vPosition. It also means it's farther
-                // away than sqrt(8) standard deviations from the mean.
-                if (A > 8.0) discard;
-                vec3 color = vColor.rgb;
-
-                // Since the rendered splat is scaled by sqrt(8), the inverse covariance matrix that is part of
-                // the gaussian formula becomes the identity matrix. We're then left with (X - mean) * (X - mean),
-                // and since 'mean' is zero, we have X * X, which is the same as A:
-                float opacity = exp(-0.5 * A) * vColor.a;
-
-                gl_FragColor = vec4(color.rgb, opacity);
-            }`;
-
-        const uniforms = {
-            'sceneCenter': {
-                'type': 'v3',
-                'value': new THREE.Vector3()
-            },
-            'fadeInComplete': {
-                'type': 'i',
-                'value': 0
-            },
-            'orthographicMode': {
-                'type': 'i',
-                'value': 0
-            },
-            'visibleRegionFadeStartRadius': {
-                'type': 'f',
-                'value': 0.0
-            },
-            'visibleRegionRadius': {
-                'type': 'f',
-                'value': 0.0
-            },
-            'currentTime': {
-                'type': 'f',
-                'value': 0.0
-            },
-            'firstRenderTime': {
-                'type': 'f',
-                'value': 0.0
-            },
-            'covariancesTexture': {
-                'type': 't',
-                'value': null
-            },
-            'centersColorsTexture': {
-                'type': 't',
-                'value': null
-            },
-            'sphericalHarmonicsTexture': {
-                'type': 't',
-                'value': null
-            },
-            'focal': {
-                'type': 'v2',
-                'value': new THREE.Vector2()
-            },
-            'orthoZoom': {
-                'type': 'f',
-                'value': 1.0
-            },
-            'inverseFocalAdjustment': {
-                'type': 'f',
-                'value': 1.0
-            },
-            'viewport': {
-                'type': 'v2',
-                'value': new THREE.Vector2()
-            },
-            'basisViewport': {
-                'type': 'v2',
-                'value': new THREE.Vector2()
-            },
-            'debugColor': {
-                'type': 'v3',
-                'value': new THREE.Color()
-            },
-            'covariancesTextureSize': {
-                'type': 'v2',
-                'value': new THREE.Vector2(1024, 1024)
-            },
-            'centersColorsTextureSize': {
-                'type': 'v2',
-                'value': new THREE.Vector2(1024, 1024)
-            },
-            'sphericalHarmonicsDegree': {
-                'type': 'i',
-                'value': maxSphericalHarmonicsDegree
-            },
-            'sphericalHarmonicsTextureSize': {
-                'type': 'v2',
-                'value': new THREE.Vector2(1024, 1024)
-            },
-            'sphericalHarmonics8BitMode': {
-                'type': 'i',
-                'value': 0
-            },
-            'splatScale': {
-                'type': 'f',
-                'value': splatScale
-            },
-            'pointCloudModeEnabled': {
-                'type': 'i',
-                'value': pointCloudModeEnabled ? 1 : 0
-            }
-        };
-
-        if (dynamicMode) {
-            uniforms['transformIndexesTexture'] = {
-                'type': 't',
-                'value': null
-            };
-            const transformMatrices = [];
-            for (let i = 0; i < Constants.MaxScenes; i++) {
-                transformMatrices.push(new THREE.Matrix4());
-            }
-            uniforms['transforms'] = {
-                'type': 'mat4',
-                'value': transformMatrices
-            };
-            uniforms['transformIndexesTextureSize'] = {
-                'type': 'v2',
-                'value': new THREE.Vector2(1024, 1024)
-            };
-        }
-
-        const material = new THREE.ShaderMaterial({
-            uniforms: uniforms,
-            vertexShader: vertexShaderSource,
-            fragmentShader: fragmentShaderSource,
-            transparent: true,
-            alphaTest: 1.0,
-            blending: THREE.NormalBlending,
-            depthTest: true,
-            depthWrite: false,
-            side: THREE.DoubleSide
-        });
-
-        return material;
-    }
-
-    /**
-     * Build the Three.js geometry that will be used to render the splats. The geometry is instanced and is made up of
-     * vertices for a single quad as well as an attribute buffer for the splat indexes.
-     * @param {number} maxSplatCount The maximum number of splats that the geometry will need to accomodate
-     * @return {THREE.InstancedBufferGeometry}
-     */
-    static buildGeomtery(maxSplatCount) {
-
-        const baseGeometry = new THREE.BufferGeometry();
-        baseGeometry.setIndex([0, 1, 2, 0, 2, 3]);
-
-        // Vertices for the instanced quad
-        const positionsArray = new Float32Array(4 * 3);
-        const positions = new THREE.BufferAttribute(positionsArray, 3);
-        baseGeometry.setAttribute('position', positions);
-        positions.setXYZ(0, -1.0, -1.0, 0.0);
-        positions.setXYZ(1, -1.0, 1.0, 0.0);
-        positions.setXYZ(2, 1.0, 1.0, 0.0);
-        positions.setXYZ(3, 1.0, -1.0, 0.0);
-        positions.needsUpdate = true;
-
-        const geometry = new THREE.InstancedBufferGeometry().copy(baseGeometry);
-
-        // Splat index buffer
-        const splatIndexArray = new Uint32Array(maxSplatCount);
-        const splatIndexes = new THREE.InstancedBufferAttribute(splatIndexArray, 1, false);
-        splatIndexes.setUsage(THREE.DynamicDrawUsage);
-        geometry.setAttribute('splatIndex', splatIndexes);
-
-        geometry.instanceCount = 0;
-
-        return geometry;
     }
 
     /**
@@ -875,9 +315,9 @@ export class SplatMesh extends THREE.Mesh {
             this.lastBuildSplatCount = 0;
             this.lastBuildMaxSplatCount = 0;
             this.disposeMeshData();
-            this.geometry = SplatMesh.buildGeomtery(maxSplatCount);
-            this.material = SplatMesh.buildMaterial(this.dynamicMode, this.antialiased, this.maxScreenSpaceSplatSize,
-                                                    this.splatScale, this.pointCloudModeEnabled, this.minSphericalHarmonicsDegree);
+            this.geometry = SplatGeometry.build(maxSplatCount);
+            this.material = SplatMaterial.build(this.dynamicMode, this.antialiased, this.maxScreenSpaceSplatSize,
+                                                this.splatScale, this.pointCloudModeEnabled, this.minSphericalHarmonicsDegree);
             const indexMaps = SplatMesh.buildSplatIndexMaps(splatBuffers);
             this.globalSplatIndexToLocalSplatIndexMap = indexMaps.localSplatIndexMap;
             this.globalSplatIndexToSceneIndexMap = indexMaps.sceneIndexMap;
@@ -1088,23 +528,20 @@ export class SplatMesh extends THREE.Mesh {
         };
 
         const covarianceCompressionLevel = this.getTargetCovarianceCompressionLevel();
-        const sphericalHarmonicsCompressionLevel = this.getTargetSphericalHarmonicsCompressionLevel();
+        const shCompressionLevel = this.getTargetSphericalHarmonicsCompressionLevel();
 
         const covariances = new Float32Array(maxSplatCount * COVARIANCES_ELEMENTS_PER_SPLAT);
         const centers = new Float32Array(maxSplatCount * 3);
         const colors = new Uint8Array(maxSplatCount * 4);
 
         let SphericalHarmonicsArrayType = Float32Array;
-        if (sphericalHarmonicsCompressionLevel === 1) SphericalHarmonicsArrayType = Uint16Array;
-        else if (sphericalHarmonicsCompressionLevel === 2) SphericalHarmonicsArrayType = Uint8Array;
-        const sphericalHarmonicsComponentCount = getSphericalHarmonicsComponentCountForDegree(this.minSphericalHarmonicsDegree);
-        let paddedSphericalHarmonicsComponentCount = sphericalHarmonicsComponentCount;
-        if (paddedSphericalHarmonicsComponentCount % 2 !== 0) paddedSphericalHarmonicsComponentCount++;
-        const sphericalHarmonics = this.minSphericalHarmonicsDegree ?
-                                   new SphericalHarmonicsArrayType(maxSplatCount * sphericalHarmonicsComponentCount) : undefined;
+        if (shCompressionLevel === 1) SphericalHarmonicsArrayType = Uint16Array;
+        else if (shCompressionLevel === 2) SphericalHarmonicsArrayType = Uint8Array;
+        const shComponentCount = getSphericalHarmonicsComponentCountForDegree(this.minSphericalHarmonicsDegree);
+        const shData = this.minSphericalHarmonicsDegree ? new SphericalHarmonicsArrayType(maxSplatCount * shComponentCount) : undefined;
 
-        this.fillSplatDataArrays(covariances, centers, colors, sphericalHarmonics, undefined,
-                                 covarianceCompressionLevel, sphericalHarmonicsCompressionLevel);
+        this.fillSplatDataArrays(covariances, centers, colors, shData, undefined,
+                                 covarianceCompressionLevel, shCompressionLevel);
 
         // set up covariances data texture
         const covTexSize = computeDataTextureSize(COVARIANCES_ELEMENTS_PER_TEXEL, 6);
@@ -1136,7 +573,7 @@ export class SplatMesh extends THREE.Mesh {
                 'covariances': covariances,
                 'centers': centers,
                 'colors': colors,
-                'sphericalHarmonics': sphericalHarmonics
+                'sphericalHarmonics': shData
             },
             'covariances': {
                 'data': paddedCovariances,
@@ -1151,39 +588,95 @@ export class SplatMesh extends THREE.Mesh {
             }
         };
 
-        if (sphericalHarmonics) {
-            const sphericalHarmonicsElementsPerTexel = 4;
-            const sphericalHarmonicsTexSize = computeDataTextureSize(sphericalHarmonicsElementsPerTexel,
-                                                                     paddedSphericalHarmonicsComponentCount);
-            const paddedSHArraySize = sphericalHarmonicsTexSize.x * sphericalHarmonicsTexSize.y * sphericalHarmonicsElementsPerTexel;
-            const paddedSHArray = new SphericalHarmonicsArrayType(paddedSHArraySize);
-            for (let c = 0; c < splatCount; c++) {
-                const srcBase = sphericalHarmonicsComponentCount * c;
-                const destBase = paddedSphericalHarmonicsComponentCount * c;
-                for (let i = 0; i < sphericalHarmonicsComponentCount; i++) {
-                    paddedSHArray[destBase + i] = sphericalHarmonics[srcBase + i];
+        if (shData) {
+            const shTextureType = shCompressionLevel === 2 ? THREE.UnsignedByteType : THREE.HalfFloatType;
+
+            let paddedSHComponentCount = shComponentCount;
+            if (paddedSHComponentCount % 2 !== 0) paddedSHComponentCount++;
+            let shElementsPerTexel = 4;
+            let shTexSize = computeDataTextureSize(shElementsPerTexel, paddedSHComponentCount);
+
+            // Based on my own observations across multiple devices, OSes and browsers, using textures that have one dimension
+            // greater than 4096 while the other is greater than or equal to 4096 causes issues (Essentially any texture larger
+            // than 4096 x 4096 (16777216) texels). Specifically it seems all texture data beyond the 4096 x 4096 texel boundary
+            // is corrupted, while data below that boundary is usable. In these cases the texture has been valid in the eyes of
+            // both Three.js and WebGL, and the texel format (RG, RGBA, etc.) has not mattered. More investigation will be needed,
+            // but for now the work-around is to split the spherical harmonics into three textures (one for each color channel).
+
+            // Use one texture for all spherical harmonics data
+            if (shTexSize.x * shTexSize.y <= 16777216) {
+                const paddedSHArraySize = shTexSize.x * shTexSize.y * shElementsPerTexel;
+                const paddedSHArray = new SphericalHarmonicsArrayType(paddedSHArraySize);
+                for (let c = 0; c < splatCount; c++) {
+                    const srcBase = shComponentCount * c;
+                    const destBase = paddedSHComponentCount * c;
+                    for (let i = 0; i < shComponentCount; i++) {
+                        paddedSHArray[destBase + i] = shData[srcBase + i];
+                    }
                 }
+
+                const shTexture = new THREE.DataTexture(paddedSHArray, shTexSize.x, shTexSize.y, THREE.RGBAFormat, shTextureType);
+                shTexture.needsUpdate = true;
+                this.material.uniforms.sphericalHarmonicsTexture.value = shTexture;
+                this.splatDataTextures['sphericalHarmonics'] = {
+                    'componentCount': shComponentCount,
+                    'paddedComponentCount': paddedSHComponentCount,
+                    'data': paddedSHArray,
+                    'textureCount': 1,
+                    'texture': shTexture,
+                    'size': shTexSize,
+                    'compressionLevel': shCompressionLevel
+                };
+            // Use three textures for spherical harmonics data, one per color channel
+            } else {
+                shElementsPerTexel = 2;
+                const shComponentCountPerChannel = shComponentCount / 3;
+                paddedSHComponentCount = shComponentCountPerChannel;
+                if (paddedSHComponentCount % 2 !== 0) paddedSHComponentCount++;
+                shTexSize = computeDataTextureSize(shElementsPerTexel, paddedSHComponentCount);
+
+                const paddedSHArraySize = shTexSize.x * shTexSize.y * shElementsPerTexel;
+                const textureUniforms = [this.material.uniforms.sphericalHarmonicsTextureR,
+                                         this.material.uniforms.sphericalHarmonicsTextureG,
+                                         this.material.uniforms.sphericalHarmonicsTextureB];
+                const paddedSHArrays = [];
+                const shTextures = [];
+                for (let t = 0; t < 3; t++) {
+                    const paddedSHArray = new SphericalHarmonicsArrayType(paddedSHArraySize);
+                    paddedSHArrays.push(paddedSHArray);
+                    for (let c = 0; c < splatCount; c++) {
+                        const srcBase = shComponentCount * c;
+                        const destBase = paddedSHComponentCount * c;
+                        if (shComponentCountPerChannel >= 3) {
+                            for (let i = 0; i < 3; i++) paddedSHArray[destBase + i] = shData[srcBase + t * 3 + i];
+                            if (shComponentCountPerChannel >= 8) {
+                                for (let i = 0; i < 5; i++) paddedSHArray[destBase + 3 + i] = shData[srcBase + 9 + t * 5 + i];
+                            }
+                        }
+                    }
+
+                    const shTexture = new THREE.DataTexture(paddedSHArray, shTexSize.x, shTexSize.y, THREE.RGFormat, shTextureType);
+                    shTextures.push(shTexture);
+                    shTexture.needsUpdate = true;
+                    textureUniforms[t].value = shTexture;
+                }
+
+                this.material.uniforms.sphericalHarmonicsMultiTextureMode.value = 1;
+                this.splatDataTextures['sphericalHarmonics'] = {
+                    'componentCount': shComponentCount,
+                    'componentCountPerChannel': shComponentCountPerChannel,
+                    'paddedComponentCount': paddedSHComponentCount,
+                    'data': paddedSHArrays,
+                    'textureCount': 3,
+                    'textures': shTextures,
+                    'size': shTexSize,
+                    'compressionLevel': shCompressionLevel
+                };
             }
 
-            const textureType = sphericalHarmonicsCompressionLevel === 2 ? THREE.UnsignedByteType : THREE.HalfFloatType;
-            const sphericalHarmonicsTex = new THREE.DataTexture(paddedSHArray, sphericalHarmonicsTexSize.x,
-                                                                sphericalHarmonicsTexSize.y, THREE.RGBAFormat, textureType);
-            sphericalHarmonicsTex.needsUpdate = true;
-            this.material.uniforms.sphericalHarmonicsTexture.value = sphericalHarmonicsTex;
-            this.material.uniforms.sphericalHarmonicsTextureSize.value.copy(sphericalHarmonicsTexSize);
-            if (sphericalHarmonicsCompressionLevel === 2) {
-                this.material.uniforms.sphericalHarmonics8BitMode.value = 1;
-            }
+            this.material.uniforms.sphericalHarmonicsTextureSize.value.copy(shTexSize);
+            this.material.uniforms.sphericalHarmonics8BitMode.value = shCompressionLevel === 2 ? 1 : 0;
             this.material.uniformsNeedUpdate = true;
-
-            this.splatDataTextures['sphericalHarmonics'] = {
-                'componentCount': sphericalHarmonicsComponentCount,
-                'paddedComponentCount': paddedSphericalHarmonicsComponentCount,
-                'data': paddedSHArray,
-                'texture': sphericalHarmonicsTex,
-                'size': sphericalHarmonicsTexSize,
-                'compressionLevel': sphericalHarmonicsCompressionLevel
-            };
         }
 
         if (this.dynamicMode) {
@@ -1211,13 +704,13 @@ export class SplatMesh extends THREE.Mesh {
         const splatCount = this.getSplatCount();
         const covarianceCompressionLevel = this.splatDataTextures['covariances'].compressionLevel;
 
-        const sphericalHarmonicsTextureDesc = this.splatDataTextures['sphericalHarmonics'];
-        const sphericalHarmonicsCompressionLevel = sphericalHarmonicsTextureDesc ? sphericalHarmonicsTextureDesc.compressionLevel : 0;
+        const shTextureDesc = this.splatDataTextures['sphericalHarmonics'];
+        const shCompressionLevel = shTextureDesc ? shTextureDesc.compressionLevel : 0;
 
         this.fillSplatDataArrays(this.splatDataTextures.baseData.covariances,
                                  this.splatDataTextures.baseData.centers, this.splatDataTextures.baseData.colors,
                                  this.splatDataTextures.baseData.sphericalHarmonics, undefined, covarianceCompressionLevel,
-                                 sphericalHarmonicsCompressionLevel, this.lastBuildSplatCount, splatCount - 1, this.lastBuildSplatCount);
+                                 shCompressionLevel, this.lastBuildSplatCount, splatCount - 1, this.lastBuildSplatCount);
 
         const covariancesTextureDescriptor = this.splatDataTextures['covariances'];
         const paddedCovariances = covariancesTextureDescriptor.data;
@@ -1233,9 +726,9 @@ export class SplatMesh extends THREE.Mesh {
             covariancesTexture.needsUpdate = true;
         } else {
             const covaranceBytesPerElement = covarianceCompressionLevel ? 2 : 4;
-            this.updateDataTexture(paddedCovariances, covariancesTextureDescriptor, covariancesTextureProps,
-                                   COVARIANCES_ELEMENTS_PER_TEXEL, COVARIANCES_ELEMENTS_PER_SPLAT, covaranceBytesPerElement,
-                                   this.lastBuildSplatCount, splatCount - 1);
+            this.updateDataTexture(paddedCovariances, covariancesTextureDescriptor.texture, covariancesTextureDescriptor.size,
+                                   covariancesTextureProps, COVARIANCES_ELEMENTS_PER_TEXEL, COVARIANCES_ELEMENTS_PER_SPLAT,
+                                   covaranceBytesPerElement, this.lastBuildSplatCount, splatCount - 1);
         }
 
         const centerColorsTextureDescriptor = this.splatDataTextures['centerColors'];
@@ -1247,35 +740,58 @@ export class SplatMesh extends THREE.Mesh {
         if (!centerColorsTextureProps || !centerColorsTextureProps.__webglTexture) {
             centerColorsTexture.needsUpdate = true;
         } else {
-            this.updateDataTexture(paddedCenterColors, centerColorsTextureDescriptor, centerColorsTextureProps,
-                                   CENTER_COLORS_ELEMENTS_PER_TEXEL, CENTER_COLORS_ELEMENTS_PER_SPLAT, 4,
+            this.updateDataTexture(paddedCenterColors, centerColorsTextureDescriptor.texture, centerColorsTextureDescriptor.size,
+                                   centerColorsTextureProps, CENTER_COLORS_ELEMENTS_PER_TEXEL, CENTER_COLORS_ELEMENTS_PER_SPLAT, 4,
                                    this.lastBuildSplatCount, splatCount - 1);
         }
 
-        if (this.splatDataTextures.baseData.sphericalHarmonics) {
-            const sphericalHarmonicsComponentCount = sphericalHarmonicsTextureDesc.componentCount;
-            const paddedSphericalHarmonicsComponentCount = sphericalHarmonicsTextureDesc.paddedComponentCount;
-            const paddedSHArray = sphericalHarmonicsTextureDesc.data;
-            for (let c = this.lastBuildSplatCount; c < splatCount; c++) {
-                const srcBase = sphericalHarmonicsComponentCount * c;
-                const destBase = paddedSphericalHarmonicsComponentCount * c;
-                for (let i = 0; i < sphericalHarmonicsComponentCount; i++) {
-                    paddedSHArray[destBase + i] = this.splatDataTextures.baseData.sphericalHarmonics[srcBase + i];
-                }
-            }
+        const shData = this.splatDataTextures.baseData.sphericalHarmonics;
+        if (shData) {
+            let shBytesPerElement = 4;
+            if (shCompressionLevel === 1) shBytesPerElement = 2;
+            else if (shCompressionLevel === 2) shBytesPerElement = 1;
 
-            const sphericalHarmonicsTex = sphericalHarmonicsTextureDesc.texture;
-            const sphericalHarmonicsTextureProps = this.renderer ? this.renderer.properties.get(sphericalHarmonicsTex) : null;
-            if (!sphericalHarmonicsTextureProps || !sphericalHarmonicsTextureProps.__webglTexture) {
-                sphericalHarmonicsTex.needsUpdate = true;
+            const updateTexture = (shTexture, shTextureSize, elementsPerTexel, paddedSHArray, paddedSHComponentCount) => {
+                const shTextureProps = this.renderer ? this.renderer.properties.get(shTexture) : null;
+                if (!shTextureProps || !shTextureProps.__webglTexture) {
+                    shTexture.needsUpdate = true;
+                } else {
+                    this.updateDataTexture(paddedSHArray, shTexture, shTextureSize, shTextureProps, elementsPerTexel,
+                                           paddedSHComponentCount, shBytesPerElement, this.lastBuildSplatCount, splatCount - 1);
+                }
+            };
+
+            const shComponentCount = shTextureDesc.componentCount;
+            const paddedSHComponentCount = shTextureDesc.paddedComponentCount;
+
+            // Update for the case of a single texture for all spherical harmonics data
+            if (shTextureDesc.textureCount === 1) {
+                const paddedSHArray = shTextureDesc.data;
+                for (let c = this.lastBuildSplatCount; c < splatCount; c++) {
+                    const srcBase = shComponentCount * c;
+                    const destBase = paddedSHComponentCount * c;
+                    for (let i = 0; i < shComponentCount; i++) {
+                        paddedSHArray[destBase + i] = shData[srcBase + i];
+                    }
+                }
+                updateTexture(shTextureDesc.texture, shTextureDesc.size, 4, paddedSHArray, paddedSHComponentCount);
+            // Update for the case of spherical harmonics data split among three textures, one for each color channel
             } else {
-                const sphericalHarmonicsElementsPerTexel = 4;
-                let sphericalHarmonicsBytesPerElement = 4;
-                if (sphericalHarmonicsCompressionLevel === 1) sphericalHarmonicsBytesPerElement = 2;
-                else if (sphericalHarmonicsCompressionLevel === 2) sphericalHarmonicsBytesPerElement = 1;
-                this.updateDataTexture(paddedSHArray, sphericalHarmonicsTextureDesc, sphericalHarmonicsTextureProps,
-                                       sphericalHarmonicsElementsPerTexel, paddedSphericalHarmonicsComponentCount,
-                                       sphericalHarmonicsBytesPerElement, this.lastBuildSplatCount, splatCount - 1);
+                const shComponentCountPerChannel = shTextureDesc.componentCountPerChannel;
+                for (let t = 0; t < 3; t++) {
+                    const paddedSHArray = shTextureDesc.data[t];
+                    for (let c = this.lastBuildSplatCount; c < splatCount; c++) {
+                        const srcBase = shComponentCount * c;
+                        const destBase = paddedSHComponentCount * c;
+                        if (shComponentCountPerChannel >= 3) {
+                            for (let i = 0; i < 3; i++) paddedSHArray[destBase + i] = shData[srcBase + t * 3 + i];
+                            if (shComponentCountPerChannel >= 8) {
+                                for (let i = 0; i < 5; i++) paddedSHArray[destBase + 3 + i] = shData[srcBase + 9 + t * 5 + i];
+                            }
+                        }
+                    }
+                    updateTexture(shTextureDesc.textures[t], shTextureDesc.size, 2, paddedSHArray, paddedSHComponentCount);
+                }
             }
         }
 
@@ -1291,8 +807,8 @@ export class SplatMesh extends THREE.Mesh {
             if (!transformIndexesTextureProps || !transformIndexesTextureProps.__webglTexture) {
                 transformIndexesTexture.needsUpdate = true;
             } else {
-                this.updateDataTexture(paddedTransformIndexes, transformIndexesTexDesc, transformIndexesTextureProps, 1, 1, 1,
-                                       this.lastBuildSplatCount, splatCount - 1);
+                this.updateDataTexture(paddedTransformIndexes, transformIndexesTexDesc.texture, transformIndexesTexDesc.szie,
+                                       transformIndexesTextureProps, 1, 1, 1, this.lastBuildSplatCount, splatCount - 1);
             }
         }
     }
@@ -1348,20 +864,19 @@ export class SplatMesh extends THREE.Mesh {
         };
     }
 
-    updateDataTexture(paddedData, textureDesc, textureProps, elementsPerTexel, elementsPerSplat, bytesPerElement, from, to) {
+    updateDataTexture(paddedData, texture, textureSize, textureProps, elementsPerTexel, elementsPerSplat, bytesPerElement, from, to) {
         const gl = this.renderer.getContext();
-        const updateRegion = SplatMesh.computeTextureUpdateRegion(from, to, textureDesc.size.x, elementsPerTexel, elementsPerSplat);
+        const updateRegion = SplatMesh.computeTextureUpdateRegion(from, to, textureSize.x, elementsPerTexel, elementsPerSplat);
         const updateElementCount = updateRegion.dataEnd - updateRegion.dataStart;
         const updateDataView = new paddedData.constructor(paddedData.buffer,
                                                           updateRegion.dataStart * bytesPerElement, updateElementCount);
         const updateHeight = updateRegion.endRow - updateRegion.startRow + 1;
-        const dataTexture = textureDesc.texture;
-        const glType = this.webGLUtils.convert(dataTexture.type);
-        const glFormat = this.webGLUtils.convert(dataTexture.format, dataTexture.colorSpace);
+        const glType = this.webGLUtils.convert(texture.type);
+        const glFormat = this.webGLUtils.convert(texture.format, texture.colorSpace);
         const currentTexture = gl.getParameter(gl.TEXTURE_BINDING_2D);
         gl.bindTexture(gl.TEXTURE_2D, textureProps.__webglTexture);
         gl.texSubImage2D(gl.TEXTURE_2D, 0, 0, updateRegion.startRow,
-                         textureDesc.size.x, updateHeight, glFormat, glType, updateDataView);
+                         textureSize.x, updateHeight, glFormat, glType, updateDataView);
         gl.bindTexture(gl.TEXTURE_2D, currentTexture);
     }
